+++ sd_dynamic_prompts/helpers.py
@@ -1,9 +1,12 @@
 from __future__ import annotations
 
 import logging
+from itertools import cycle, islice, product
 from pathlib import Path
 
 from dynamicprompts.generators.promptgenerator import PromptGenerator
+
+from sd_dynamic_prompts.paths import get_magicprompt_models_txt_path
 
 logger = logging.getLogger(__name__)
 
@@ -14,7 +17,7 @@ def get_seeds(
     use_fixed_seed,
     is_combinatorial=False,
     combinatorial_batches=1,
-):
+) -> tuple[list[int], list[int]]:
     if p.subseed_strength != 0:
         seed = int(p.all_seeds[0])
         subseed = int(p.all_subseeds[0])
@@ -47,25 +50,22 @@ def should_freeze_prompt(p):
     return p.subseed_strength > 0
 
 
-def load_magicprompt_models(modelfile: str) -> list[str]:
+def load_magicprompt_models(models_file: Path | None = None) -> list[str]:
+    if not models_file:
+        models_file = get_magicprompt_models_txt_path()
     try:
-        models = []
-        with open(modelfile) as f:
-            for line in f:
-                # ignore comments and empty lines
-                line = line.split("#")[0].strip()
-                if line:
-                    models.append(line)
-        return models
+        # ignore empty lines
+        return [
+            model
+            for model in (
+                line.partition("#")[0].strip()
+                for line in models_file.read_text().splitlines()
+            )
+            if model
+        ]
     except FileNotFoundError:
-        logger.warning(f"Could not find magicprompts config file at {modelfile}")
+        logger.warning(f"Could not find magicprompts config file at {models_file}")
         return []
-
-
-def get_magicmodels_path(base_dir: str) -> str:
-    magicprompt_models_path = Path(base_dir / "config" / "magicprompt_models.txt")
-
-    return magicprompt_models_path
 
 
 def generate_prompts(
@@ -74,7 +74,7 @@ def generate_prompts(
     prompt: str,
     negative_prompt: str | None,
     num_prompts: int,
-    seeds: list[int],
+    seeds: list[int] | None,
 ) -> tuple[list[str], list[str]]:
     """
     Generate positive and negative prompts.
@@ -100,9 +100,50 @@ def generate_prompts(
         seeds=negative_seeds,
     ) or [""]
 
-    if len(all_negative_prompts) < len(all_prompts):
-        factor = len(all_prompts) // len(all_negative_prompts) + 1
-        all_negative_prompts = all_negative_prompts * factor
-    all_negative_prompts = all_negative_prompts[: len(all_prompts)]
+    if num_prompts is None:
+        return generate_prompt_cross_product(all_prompts, all_negative_prompts)
 
-    return all_prompts, all_negative_prompts
+    return all_prompts, repeat_iterable_to_length(all_negative_prompts, num_prompts)
+
+
+def generate_prompt_cross_product(
+    prompts: list[str],
+    negative_prompts: list[str],
+) -> tuple[list[str], list[str]]:
+    """
+    Create a cross product of all the items in `prompts` and `negative_prompts`.
+    Return the positive prompts and negative prompts in two separate lists
+
+    Parameters:
+    - prompts: List of prompts
+    - negative_prompts: List of negative prompts
+
+    Returns:
+    - Tuple containing list of positive and negative prompts
+    """
+    if not (prompts and negative_prompts):
+        return [], []
+
+    # noqa to remain compatible with python 3.9, see issue #601
+    new_positive_prompts, new_negative_prompts = zip(
+        *product(prompts, negative_prompts),  # noqa: B905
+    )
+    return list(new_positive_prompts), list(new_negative_prompts)
+
+
+def repeat_iterable_to_length(iterable, length: int) -> list:
+    """Repeat an iterable to a given length.
+
+    If the iterable is shorter than the desired length, it will be repeated
+    until it is long enough. If it is longer than the desired length, it will
+    be truncated.
+
+    Args:
+        iterable (Iterable): The iterable to repeat.
+        length (int): The desired length of the iterable.
+
+    Returns:
+        list: The repeated iterable.
+
+    """
+    return list(islice(cycle(iterable), length))