+++ sd_dynamic_prompts/settings.py
@@ -1,11 +1,7 @@
-from pathlib import Path
-
 import gradio as gr
-from modules import scripts, shared
+from modules import shared
 
-from sd_dynamic_prompts.helpers import get_magicmodels_path, load_magicprompt_models
-
-base_dir = Path(scripts.basedir())
+from sd_dynamic_prompts.helpers import load_magicprompt_models
 
 
 def on_ui_settings():
@@ -76,7 +72,34 @@ def on_ui_settings():
         ),
     )
 
-    magic_models = load_magicprompt_models(get_magicmodels_path(base_dir))
+    shared.opts.add_option(
+        key="dp_wildcard_manager_no_dedupe",
+        info=shared.OptionInfo(
+            False,
+            label="Disable de-duplication of wildcards before processing.",
+            section=section,
+        ),
+    )
+
+    shared.opts.add_option(
+        key="dp_wildcard_manager_no_sort",
+        info=shared.OptionInfo(
+            False,
+            label="Disable sorting of wildcards before processing.",
+            section=section,
+        ),
+    )
+
+    shared.opts.add_option(
+        key="dp_wildcard_manager_shuffle",
+        info=shared.OptionInfo(
+            False,
+            label="Shuffle wildcards before use for more random outputs",
+            section=section,
+        ),
+    )
+
+    magic_models = load_magicprompt_models()
     shared.opts.add_option(
         key="dp_magicprompt_default_model",
         info=shared.OptionInfo(