+++ docs/CHANGELOG.md
@@ -1,3 +1,11 @@
+- 2.16.3 Refactoring and fixing [#601](https://github.com/adieyal/sd-dynamicprompts/issues/601)
+- 2.16.2 Refactor paths and magic prompt models file read
+- 2.16.1 Only using cross product when num_prompts is not provided
+- 2.16.0 Added cross product of positive and negative prompts
+- 2.15.0 Added Wildcard Manager Search
+- 2.14.0 Added configuration option to shuffle wildcards for increased randomness in combinatorial mode.
+- 2.13.0 Added configuration options to prevent the wildcard manager from sorting and deduplicating wildcard files
+- 2.12.6  Bumped dynamicprompts version to 0.27.1. Late loading of jinja2 to work around issues with incompatible versions of Jinja2 being installed. Also improved dependency management
 - 2.12.5  Updates to support hires updates in auto1111. See [#498](https://github.com/adieyal/sd-dynamic-prompts/pull/498). Also manually updates PYTHONPATH in install.py if the extension's package is not found. [#496](https://github.com/adieyal/sd-dynamic-prompts/pull/496)
 - 2.12.4  Honouring --skip-install command line argument for faster load times. Also reverts change to extension-specific logging configuration - see [53e69fe](https://github.com/adieyal/sd-dynamic-prompts/commit/53e69feda77954b86b4b0bae5623bd85009826b3)
 - 2.12.3  Minor change that does removes the "Promt matrix..." log when there is a single prompt